Manual roll of Dart from 5ca1f32905 to 551e405d0b (flutter/engine#20758)
dart-lang/sdk@551e405d0b Version 2.10.0-57.0.dev
 dart-lang/sdk@ce494e7910 Add test for inference solving T? = dynamic.
 dart-lang/sdk@0e1d32f10c [fuchsia][time] Use C11 timespec APIs for UTC time
 dart-lang/sdk@08dd4b1faa Null-safe numbers: implement static typing rules, hook up for binary operators.
 dart-lang/sdk@9be24ce946 [benchmark] Import MD5/SHA1/SHA256/SkeletalAnimation/SkeletalAnimationSIMD benchmarks from benchmarks-internal
 dart-lang/sdk@675891ae0c Issue 43152. Use existing LineInfo in newDiagnosticMessage().
 dart-lang/sdk@e868d3076c Issue 43151. Fix crash when a field shadows a super-interface method.

+++ engine/src/flutter/lib/web_ui/lib/src/engine/text/ruler.dart
@@ -794,9 +794,9 @@ class ParagraphRuler {
       }
 
       boxes.add(ui.TextBox.fromLTRBD(
-        rect.left + alignOffset as double,
+        rect.left + alignOffset,
         rect.top as double,
-        rect.right + alignOffset as double,
+        rect.right + alignOffset,
         rect.bottom as double,
         textDirection,
       ));
